backend/requirements.txt
Django>=5.2,<5.3
django-jazzmin>=3.0,<4
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.8,<4.9
django-filter>=25.1,<25.3
python-dotenv>=1.1,<1.2
psycopg[binary]>=3.2.13,<3.4
stripe>=14,<15
Pillow>=12,<13
